requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
asgiref==3.8.1
    # via
    #   django
    #   django-cors-headers
django==5.0.6
    # via
    #   -r requirements.in
    #   django-cors-headers
    #   django-filter
    #   djangorestframework
django-cors-headers==4.4.0
    # via -r requirements.in
django-filter==24.2
    # via -r requirements.in
djangorestframework==3.15.2
    # via -r requirements.in
pillow==10.3.0
    # via -r requirements.in
psycopg2-binary==2.9.9
    # via -r requirements.in
sqlparse==0.5.0
    # via django
